リファクタリング
実装内容：
・追加ボタンのタイプをsubmitからbuttonに変更。目的は追加ボタンクリック時のエラーを解消するため。
・jsonファイルの情報を読みこむ処理を別のjsファイルに記述した。
・追加タスクの情報をモックサーバーに登録後に、jsonファイルを読み込んでタスクを表示。
